Sign out bug fixed
Every time the sign out button was pressed the app would crash. It took
two days to find the problem. The error was caused by Log.d("MyTag",
"onAuthStateChanged:signed_in:" + user.getUid()); were user.getUid() was
looking for an id that did not exist when the sign out button was
pressed. this is solved