After meeting updates
I got the name EditText fieldsto capitalise the first letter and the DOB
EditText will now change over once the right length of numbers are
entered. I spent some time trying to get the password to show what is
missing but I cant spend too much time on it so I will have to come back
to it.