Fixed the verification feature
I thought the taxi verification was finally fixed today by using a
timer. What I noticed later was that it was still showing that the taxi
is not registered even when it was. This was solved after a lot of try
and error by simply adding an if statement to onChildAdded method that
if it reads data then cancel the timer. Otherwise the timer is ok to
proceed